requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
executing==2.0.1
fastapi==0.110.2
h11==0.14.0
httptools==0.6.1
icecream==2.1.3
idna==3.7
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
python-dotenv==1.0.1
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.1
sslib==0.2.0
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
